requirements.txt
aiohttp==3.6.2
aiohttp-cors==0.7.0
aioredis==1.3.1
altgraph==0.15
amqp==2.5.2
argh==0.26.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
Babel==2.7.0
billiard==3.6.1.0
boxsdk==2.6.1
celery==4.3.0
certifi==2018.4.16
cffi==1.11.5
chainmap==1.0.2
chardet==3.0.4
Click==7.0
cryptography==2.8
dis3==0.1.2
enum34==1.1.6
et-xmlfile==1.0.1
flower==0.9.3
future==0.16.0
hiredis==1.0.1
idna==2.6
importlib-metadata==1.3.0
ipaddress==1.0.22
jdcal==1.4.1
kombu==4.6.7
lxml==4.4.2
macholib==1.9
more-itertools==8.0.2
multidict==4.7.1
ndg-httpsclient==0.5.1
openpyxl==3.0.2
pathtools==0.1.2
pefile==2017.11.5
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.18
PyInstaller==3.3.1
PyJWT==1.6.1
pyOpenSSL==19.1.0
pytz==2018.4
PyYAML==5.2
redis==3.3.11
requests==2.18.4
requests-toolbelt==0.8.0
service-identity==18.1.0
six==1.10.0
slack==0.0.2
tornado==5.1.1
tzlocal==1.5.1
urllib3==1.22
vine==1.3.0
watchdog==0.9.0
wrapt==1.11.2
yarl==1.4.2
zipp==0.6.0
